Introduce UnresolvedNdkCxxPlatform
Summary:
Following the pattern introduced by UnresolvedCxxPlatform, introduce an
"unresolved" version of NdkCxxPlatform. This allows descriptions to
refer to this in the unresolved form prior to graph creation so that
they can appropriately get parse time deps.

This will enable an ndk_toolchain rule that will allow us to fully
specify the ndk toolchain in the repo.

Updates a couple rules to use this correctly. cxx rules use it correctly
via the corresponding UnresolvedCxxPlatform.

Reviewed By: sbalabanov

fbshipit-source-id: 72d3d48375