Introduce ndk_toolchain
Summary:
This adds the buckconfig support for specifying that an ndk platform is
configured by a ndk_toolchain target. The implementation side of the change is
pretty straightforward, the hard work there is done in previous diffs.

This follows the pattern of cxx_toolchain rule (and steals some of its test
support).

Reviewed By: sbalabanov

fbshipit-source-id: 40f43c41e5