Fix Building with Xcode 10.2 beta 4
Summary:
Xcode 10.2 beta 4 changed the lipo binary to error if you try extracting a slice from a binary that only has one slice (even if the binary contains that one slice, i.e. x86_64 from an x86_64 binary will now error). This breaks copying the swift swift stdlib as its x86_64 only.

To fix this, we now check each stdlib binary to see if it has available architectures to extract, and if it doesn't, we just copy them as-is. (which probably won't change anytime soon)

Reviewed By: ryu2

fbshipit-source-id: c57d5c331d