Fixed caching issue when switching xcode_public_headers_symlinks on and off
Summary:
Changing the value of the setting xcode_public_headers_symlinks for a cxx_library rule wouldn't invalidate properly the cache item.

The build would share the same cache item with xcode_public_headers_symlinks on and off.
The input based rulekey ended up being the same.

This diff introduce an additional value in the rulekey: it adds the name of the header symlink tree class. It helps differentiate the cache entries between headermap with symlinks and headermap without symlinks.

Reviewed By: cjhopman

fbshipit-source-id: d7364d51c9