Allow sending target names to `buck cache`
Summary:
If we want to use the target name in the cache request, then we need to make sure we send the target name as often as possible. Right now, when you invoke `buck cache`, there is no way to associate a rulekey with a build target - it always uses null.

In this diff, I make it so that `buck cache` supports sending a target name along with a rulekey.

Reviewed By: styurin

fbshipit-source-id: 7c8e40cf2e