Initial work on multitenant Buck graph service.
Summary:
This lays the groundwork for setting up an efficient index for build
rules and their dependencies over a linear chain of commits.

As demonstrated by `IndexTest`, the idea is that changes to the build
graph are added to the `Index` for each commit in a chain. For the
first commit, the entire build graph would have to be parsed and the
initial `Changes` object would have a `addedBuildPackages` list that
contained the entire build graph (its `modifiedBuildPackages` and
`removedBuildPackages` would be empty).

For subsequent commits, it is a matter of inspecting the files changed
in the commit:
- If a new `BUCK` file is added, `buck audit rules` for that file should be
  run and the result should be reflected in `addedBuildPackages`.
- If a `BUCK` file is removed, that should be reflected in `removedBuildPackages`.
- If other files are added or removed, then the nearest `BUCK` file should
  be found and `buck audit rules` should be re-run for each, the results of
  which should be added to `modifiedBuildPackages`. The `Index` will take
  responsibility for determining whether the build package was actually
  modified or not.

(Note when adding or removing a `BUCK` file, the parent build package
must also be checked as it could end up in the `modifiedBuildPackages`.)

Note this does not currently take other changes that could affect the build
graph into account, such as modifying `.buckconfig` or a `.bzl` file.
Those cases will be handled in the future.

Reviewed By: sbalabanov

fbshipit-source-id: 5281cdb147