Move checking for compatible targets from TargetNode creation to target graph creation
Summary:
The current model with failing `TargetNode` creation can result in nodes created in some situations and failing creation in other. Since we use cache to store created objects some execution paths may result in succeeding when a previous operation resulted in creating a node without checking for compatibility.

This change moves checking for compatibility to a later stage when a target graph is created.

Reviewed By: bobyangyf

fbshipit-source-id: 201d9757c1