Refactor filtering incompatible targets to happen after TargetNodeSpec resolution
Summary:
This change modifies the logic to move filtering by target platform compatibility to a later stage, after target node specs are processed.

The main drawback of this approach is the need to access `TargetNode` while constructing `BuildTarget`. This makes logic around `TargetNodeSpecResolver` more complicated then it needs to be.

Reviewed By: sbalabanov

fbshipit-source-id: 07abcba647