Convert CalculateSourceAbi to MBR
Summary:
This is a pretty straightforward one, because it's mostly just
implemented by the same thing as DefaultJavaLibrary which has already
been converted.

This is an important one since the with-deps version runs in a pipeline
with the corresponding DefaultJavaLibrary. That means that the pipeline
can get started locally disabling RE for the DJL.

Reviewed By: jtorkkola

fbshipit-source-id: 1a86f64799