Pass target configuration in parse pipelines
Summary:
The main change is extending `ParsePipeline` API to pass `TargetConfiguration` to methods that produce result and don't have a way to deduce `TargetConfiguration`. For example, getting all target nodes from a build file requires a target configuration while getting a node for a build target does not because `BuildTarget` provides access to `TargetConfiguration`.

The rest of the changes is plumbing to pass `TargetConfiguration` to the actual calls of those methods.

Reviewed By: bobyangyf

fbshipit-source-id: 735f671327