Fixed location of the generated workspace
Summary:
Before the diff, when generating a project of a target in a cell, it would put the workspace in the folder
in the main cell and not in the specific cell.

```
$ buck project foobar//path/to:target
```

It would write the workspace to [main-cell-path]/path/to (a) instead of [foobar-cell-path]/path/to (b).
The expected location is (b). Especially, since the project for [foobar-cell-path]/path/to is generated in that folder.

This diff makes sure that the workspace is generated at the correct location when the target is in a cell.

Reviewed By: williamtwilson

fbshipit-source-id: e98c4f159a