Make StripLinkable filesystem match the underlying rule/target
Summary:
When AndroidBinary creates strip rules for its native libraries, it creates
them using the buildtargets of the libraries (so that multiple
android_binary can share the same strip rule), but it uses its own
filesystem. This causes cross-cell strip rules to be created with a
filesystem that doesn't match their own buildtarget...

We'll make ModernBuildRule enforce that these match so we can catch
other instances as we migrate (which hopefully don't exist).

Reviewed By: rajyengi

fbshipit-source-id: 1a65cc943a