Add ApplicationIdentifier to Info.plist
Summary:
We need to add the AppID as part of the buck build step because Info.plist needs to be updated before the bundle is signed.
I created the AppID value getting the prefix from the provisioning profile (first element) and the BundleID from the Info.plist (we cannot use the provisioning profile because it can be generic, i.e `com.facebook.*`).

Reviewed By: ryu2, dinhviethoa

fbshipit-source-id: e8a60d3da4