Rewrite extension loading mechanism to be iterative
Summary:
Rewrite extension loading mechanism to be iterative

Current implementation relies on recursion to load extensions: if an extension loads other extensions, a recursive call
is made to load each extension in turn.  This is not very efficient, and becomes less efficient as the complexity of the macros grows.  Modify extension loading mechanism to use a work queue of extensions to load.

Reviewed By: sbalabanov

fbshipit-source-id: 473e3703ae