Add PerfRuleKeyCommand
Summary:
Buck doesn't really have any way to test components in isolation in
real-world environments (and has very little of it even in synthetic
environments). This change adds an approach to testing such things in
the real world. It adds a command that can be invoked like
`buck perf rk --key-type input //src/com/facebook/buck/cli/...`
that will try to test the performance of just computing input-based keys (or
other types of keys) on the graph of whatever targets are passed.

This will enable all kinds of cool things: evaluating perf of changes,
more reliable regression detection, easier regression reproduction,
targeted profiling, etc.

This actually adds a couple things

`buck perf <...>` PerfCommand is simply a container command for targeted
perf testing of buck components.

AbstractPerfCommand - this is the base class for the actual perf
commands. It has a couple arguments --repeat=n --ignore-first=x
--stop-after-seconds=yyy that will cause us to run the perf test n
times, ignore the first x for statistics and stop early if we exceed yyy
seconds. It will ask the implementation to do some setup and then run
the perf tests. It'll print some progress information and then at the
end print the mean and 95% interval for timings.

`buck perf rk --key-type <type> xxx` PerfRuleKeyCommand - this command
can be used to test the performance of computing each of the different
rulekey types. It injects fake filesystems that don't actually access
the filesystem for hashes and other functions (we could maybe make this
configurable which could help with checking perf of eden
integrations for example).

For depfile/manifest as these computations are primarily implemented by
the rules themselves and because the rules have been written to read
from their own outputs and do other bad stuff, it'll only work after a
full --deep build has been performed. In addition, it requires a couple
of --unsafe-<foo> flags because we need to do some additional setup but
we can't afford to actually do a build. This should be fine, but I think
a scary flag name is good.

Reviewed By: philipjameson

fbshipit-source-id: e23e542954