Update Index to store all attributes of a rule, not just deps.
Summary:
Previously, a key data structure inside `Index` was of type:

```
GenerationMap<BuildTargetId, BuildTargetSet, BuildTargetId>({ it })
```

and now it is:

```
GenerationMap<BuildTargetId, InternalRawBuildRule, BuildTargetId>({ it })
```

where `InternalRawBuildRule` is:

```
data class InternalRawBuildRule(val targetNode: RawTargetNode, val deps: BuildTargetSet)
```

This means that for an `(UnconfiguredBuildTarget, generation)` pair,
we can efficiently lookup the `RawTargetNode` and `deps` for the build
rule at that point in time. This will allow us to support richer queries
over the index.

Reviewed By: sbalabanov

fbshipit-source-id: 31f91c9e84