Load extensions once
Summary:
Ensure each extension is loaded only once.

Change extension data cache to use path as its key.  The use of LoadImport as the cache key is incorrect: it results in a different hash being produced for each time an extension is included from a different file (or even each time the load() statement differs).  Now, we use path as the key and ensure that every time we lookup extension data, we update the returned value with the correct import string.

Reviewed By: styurin

fbshipit-source-id: 3e7f8821d9