Provide default implementation of compareTo() in QueryTarget.
Summary:
I am looking into providing an implementation of `QueryEnvironment`
that operates on a `com.facebook.buck.multitenant.service.Index`.
As such, I was looking into what it means to be a `QueryTarget`, and
it's...a bit weird.

Prior to this change, it was clearly a marker interface, as it defined
no methods, but it did claim to implement `Comparable`. This is a bit
odd because how are arbitrary implementations of an interface supposed
to know how they compare to one another?

Looking at the two implementations (`QueryBuildTarget` and
`QueryFileTarget`), it turns out that they both implemented
`compareTo()` based on their string representation, so this diff moves
that logic up into `QueryTarget` itself to make that explicit.

Since I was updating the Javadoc, anyway, I wanted to call attention to
the fact that the use of a marker interface in the
`com.facebook.buck.query` package is misguided. That is,
`QueryEnvironment` frequently appears in method signatures where the
underlying implementation accepts only a specific implementation of
`QueryEnvironment`.

For example, `BuckQueryEnvironment` leans heavily on its internal
`getNode(QueryTarget target)` method that throws if `target` is not
an instance of `QueryBuildTarget`. To that end, it would be better if
method signatures could be tightened up to eliminate the possibility of
these runtime typecheck failures. Ideally (as illustrated in the
Javadoc), algebraic data types could be used to make it impossible to
represent invalid state that is possible in today's codebase.

In particular, `BuckQueryEnvironment`

Reviewed By: bobyangyf

fbshipit-source-id: d1f3b11cd1