Introduce cxx.binary_extension config setting
Summary:
There're already settings for static and shared libraries and object files but setting for binary extension was missing. That's needed for building android binaries on Windows with correct filename.

If `cxx.binary_extension` is not set, default extension for the platform will be used. If it's set to empty string no extension will be added.

Reviewed By: styurin

fbshipit-source-id: f1aec75c2c