Don't install APK-under-test if it is already present
Summary:
The APK-under-test can be quite big and it is wasteful to install it again if it is already present. Therefore before installing we check to see if we have an APK installed with the same package as the target package, and if we do, we check the MD5 hash of the installed APK against the APK we are going to install. We doing this before we sync to the device and before we actually install.

The time for looking for the APK is small (a few hundred milliseconds or less), whereas the time to do the MD5 hash is large (~1s per APK). However, in most cases if we find an APK we'd expect it to be the same as what we want to install so this seems reasonable.

Reviewed By: asp2insp

fbshipit-source-id: f0315d5932