Restrict QueryEnvironment.getTargetKind() to QueryBuildTarget.
Summary:
Many methods in `QueryEnvironment` claim to accept a `QueryTarget`,
but actually accept only a `QueryBuildTarget`. This is a first step in
an attempt to make all of the methods of `QueryEnvironment` more
specific.

I think that, in the limit, we probably want to replace
`QueryBuildTarget` with a generic type `T` to make it easier to use
with `com.facebook.buck.multitenant.service.Index`.

Reviewed By: bobyangyf

fbshipit-source-id: 45d6823cf7