Migrate more methods to QueryBuildTarget.
Summary:
In `QueryEnvironment`, migrate `getReverseDeps()` and
`getTransitiveClosure()` to work with `QueryBuildTarget` instead of
`QueryTarget`.

Admittedly, this is really gross. I'm going to keep pulling on this
thread for a bit to see what happens, but I expect I will have to do
quite a bit of cleanup before this is ready for review.

Reviewed By: sbalabanov

fbshipit-source-id: 8193b6edf6