Add simple (non-robust) utility for dumping heaps
Summary:
I found myself running perf commands and seeing how long an iteration
took and then running in another terminal `sleep xxx; jmap <...>` to try
to capture a heap dump near the end of an iteration (to capture all the
cached state). That's really dumb.

This adds a really simple mechanism for programatically capturing a heap
dump. It may not be robust enough for real production use, but within
perf commands it's great.

Adds a flag to `perf mbr prepare-remote-execution` that will capture a
heap dump at the end of the first iteration.

Reviewed By: rajyengi

fbshipit-source-id: 3da5251abf