Cache NodeData directly on the MerkleTreeNode
Summary:
We're currently holding a separate map to cache NodeData objects when
they are requested. That's silly, these are all internal data
structures let's just cache it directly on the MerkleTreeNode.

Since callers are caching the MerkleTreeNodes that they care about, this
will also allow us to easily change the interner to be weak (since then
we don't have our internal NodeData cache holding on to references).

Reviewed By: rajyengi

fbshipit-source-id: 8d1081a153