Make the nodeInterner weak.
Summary:
For efficient merging of merkle trees, we expect the caller to be
holding onto references to the nodes that it cares about and so we can
just weakly intern them. Then, if the caller has some nodes that it
doesn't expect to reuse we'll just naturally drop our cache of them when
the caller drops its references to them.

Reviewed By: rajyengi

fbshipit-source-id: 52739f4886