Don't convert Paths to Strings for map keys
Summary:
Converting BuckUnixPath to a String is super-expensive. For part of
Manifest manipulation, we need to stick some paths into a map as keys.
Just using the Path is much, much more efficient than converting it to a
String.

In real-world data, these manifest operations account for somewhere between
2-5% of total amortized time. In synthetic tests, this change is a 3-4x
speedup.

Reviewed By: rajyengi

fbshipit-source-id: 8eed6207e7