Fix BuckGlobalState to not keep empty Watchman cursors
Summary:
If Watchman fails to start then BuckGlobalState is still created and kept. Along with state, we keep Watchman cursors, one per cell. If Watchman restores, we keep reusing that state but in fact no Watchman updates are seen because cursors map is empty.
This change makes state to recycle when Watchman status change, one way or another.

Reviewed By: bobyangyf

fbshipit-source-id: 1f098674d2