Make total input size an attribute of the merkle tree nodes
Summary:
This is a minor optimization such that we don't recompute these, but
that's not the point. We'd like to change the requireddata of actioninfo
only be lazily computed based on what we know or don't know is in the CAS
already. For us to do that, we need to compute the inputs size
differently.

Reviewed By: rajyengi

fbshipit-source-id: ff4ee57df5