Optimize serialization+inputs derivation of JavaDependencyInfos
Summary:
A lot of the RuleKey-based computations (rulekey computations,
dependency derivation, serialization, inputs derivation) are optimized
by caching the derived values for all AddsToRuleKey objects.

These fields in JarBuildStepsFactory are not shared with anything else
and so don't benefit from that optimization and there's a ton of these
in large build graphs.

We've already had to do custom rulekey handling for these to not destroy
the performance of rk calculation and now we just extend that handling
to apply to serialization and inputs-calculation.  This is a significant
win for both performance and memory use.

Reviewed By: rajyengi

fbshipit-source-id: 572d88f9df