Drop serialized object data after upload
Summary:
We're currently caching the serialized object data in memory and this
can end up being quite large. We really only need it once (though we
don't know exactly which consumer is going to be the one that
uploads it). This adds a mechanism where we will drop the serialized
data after it is used. Since we might get multiple concurrent requests
for the data before it's uploaded, we stash it in a WeakReference after
the first request (and so it just naturally gets dropped whenever the
requestor actually uploads it).

To handle the rare edge cases that might actually need the serialized
data multiple times, we just support explicitly reserializing it.

Reviewed By: rajyengi

fbshipit-source-id: 215028b27b