Consolidate grammar docs in tools/psi-grammar/docs
Summary:
Put copies of all of the grammar-related documentation in a single
place to make it clearer *which* versions of reference material
were used to generate the buck grammar.  These materials have been
periodically updated/rehomed, so hopefully keeping a code locally
will be useful for future developers to check for changes in the specs.

Reviewed By: asp2insp

fbshipit-source-id: bd647afa7c