Skip a ton of work once an item has been uploaded to the cache
Summary:
We're currently creating these large maps of inputs for things that
we've already uploaded to the cache. This is a ton of unnecessary work.

Expose from the cache a `containsDigest()` method and use this in
prepareRemoteExecution to construct only the minimal required data list
that may still be needed to upload.

Reviewed By: bobyangyf

fbshipit-source-id: d94a0ab9b9