Generify QueryBuildTarget in QueryEnvironment.
Summary:
Prior to this change, a `QueryEnvironment` primarily exposed the information in
its environment in the form of `QueryBuildTarget` objects. The network of
`QueryFunction` and `Argument` objects it collaborated with were also built
around the assumption of working with `QueryBuildTarget` objects.

For the graph service, we want to provide an alternate `QueryEnvironment`
implementation that pulls its data from a `com.facebook.buck.multitenant.service.Index`.
Note that the `Index` works with `RawTargetNode`s and `UnconfiguredBuildTarget`s,
so it does not easily lend itself to operations in terms of `QueryBuildTarget`s.

To that end, this diff takes all references to `QueryBuildTarget` in `QueryEnvironment`
and replaces them with a generic `NODE_TYPE`. This required making key
collaborators of `QueryEnvironment` generic, as well:

- Because a `QueryExpression` is evaluated in terms of a `QueryEnvironment`,
  it also needs to be parameterized by `NODE_TYPE` to ensure the environment used
  to evaluate it matches the expression.
- The most common type of `QueryExpression` is a `QueryFunction`, so it must
  also be parameterized by `NODE_TYPE`. Note that `QueryFunction` has two generics
  associated with it: `NODE_TYPE` and `OUTPUT_TYPE`. In practice, most
  `QueryFunction`s today use `QueryBuildTarget` for both, but a `QueryFunction`
  may very well take an environment based on `QueryBuildTarget`, but return
  values of type `QueryFileTarget`, so they must be able to be considered separately.
- Because `QueryFunction.eval()` takes a list of `QueryEnvironment.Argument`
  objects and `Argument` may wrap a `QueryExpression`, `Argument` also had to
  be expanded to be parameterized by `NODE_TYPE`.
- Relatedly, `QueryEvaluator` is also a parameter to `QueryFunction.eval()` that
  operates on a `QueryExpression` and `QueryEnvironment` itself, so it also had
  to be changed to be parameterized in terms of `NODE_TYPE`.

In the course of working on this diff, I had trouble with `TargetLiteral` and immutables
because the generated `TargetLiteral.of()` method did not seem to be generated in
the way that I needed (or at least, I was not able to use `TargetLiteral::of` in the way
I was before, though I suppose this could have only been an issue with IntelliJ's
support for incremental builds in the context of annotation processors). I decided to
delete `AbstractTargetLiteral` and implement `TargetLiteral` directly since it is basically
a glorified wrapper for `String`.

(Incidentally, the use of `(prehash = true)` on `AbstractTargetLiteral` seemed
unnecessary to me because `String` already caches its hash code, so it seemed
gratuitous to have a separate `hashCode` field on `TargetLiteral`.)

Admittedly, many of the implementations of `QueryFunction` are hardcoded to
use `QueryBuildTarget` as both the `NODE_TYPE` and `OUTPUT_TYPE`.
(If you look at `BuckQueryEnvironment.QUERY_FUNCTIONS`, you can see that only
a handful of functions had to be updated to specify the generic.) I expect this
to change when I introduce my own `QueryEnvironment` implementation that wants
to reuse as many of these functions as possible.

Reviewed By: sbalabanov

fbshipit-source-id: 34765b9daa