Move QueryTargetAccessor into the com.facebook.buck.rules.query package.
Summary:
`QueryTargetAccessor` was the only class in the `com.facebook.buck.query`
package with a dependency on
`//src/com/facebook/buck/core/model/targetgraph:targetgraph`, which
caused it to have unexpected transitive dependencies on things like
`//third-party/java/android:common` and `//third-party/java/android:ddmlib`.

The primary users of `QueryTargetAccessor` are:
* `com.facebook.buck.rules.query.GraphEnhancementQueryEnvironment`
* `com.facebook.buck.cli.BuckQueryEnvironment`

To that end, `com.facebook.buck.rules.query` seemed like a good home
for `QueryTargetAccessor` to help keep the `com.facebook.buck.query`
leaner.

This will help clear the way for a new implementation of
`QueryEnvironment`, `MultitenantQueryEnvironment`, that has a reasonable
set of transitive dependencies.

Reviewed By: sbalabanov

fbshipit-source-id: d641767715