Push changed modules to disk rather than passing via an Intent
Summary: Previously, we were passing all the changed classes in the module via an Intent, but this doesn't work in all cases because adb shell has a command length limit that can be exceeded when we are hotswapping a large module (see e.g. https://fburl.com/eptdqbhf). Instead, we can just push a file that we then read.

Reviewed By: asp2insp

fbshipit-source-id: cd34a1603d