Add a directory to external_runner_specs.json
Summary:
Right now the contract that Buck has with external runners is "we give you a command and some metadata, you run the test".

The "command" part is tricky though, as it's built from both Buck logic and macro logic. Buck itself can be smart about only using absolute paths in the command, but macro logic doesn't have that option and so only uses relative paths.

Using relative paths means that the command is inherently tied to the directory it is run in. Buck's internal runner normally runs tests from the cell root of whichever cell owns the test, so this diff adds an extra field to the external_runner_specs.json file which tells the external runner which directory the runner should run the test from.

Reviewed By: asp2insp

fbshipit-source-id: 3a61fa0a5e