Rewrite parallel action graph to be futures-based
Summary:
For the ForkJoinPool-based implementation it has been very difficult to
 understand how the underlying Java implementation behaves in
different scenarios and to coerce it into behaving the way that we want.

This changes the parallel action graph to instead just use futures. It can
only parallelize based on calls to the Parallelizer or to
 requireAllRules/computeAllIfAbsent, but that's not really any different.

The Futures are not exposed to callers, instead it acts pretty much how
we wanted FJP to act: when it knows it has extra work to do, it forks it
off to possibly run on other threads and then it tries to complete the work
itself.

Reviewed By: bobyangyf

fbshipit-source-id: 3e4a36a51b