Implement buck fetch without hooking into internals
Summary:
Currently, FetchCommand overrides/injects a couple of core pieces of
core buck behavior: (1) it overrides the
TargetNodeToBuildRuleTransformer and (2) it uses
TargetNode.copyWithDescription.

The support for injecting/changing such behavior is super-powerful but
it also makes it really hard for us to change or even understand things.

This changes buck fetch to be implemented mostly on the normal action
graph generation. It simply (1) injects a config override to allow
downloading and (2) it traverses the AG to find the downloadable rules
and only build those.

This is the only use of copyNodeWithDescription and so that is deleted.

Reviewed By: jtorkkola

fbshipit-source-id: 18fe220780