Fix running when exopackage is present
Summary:
Android instrumentation tests which were exopackage enabled were failing when run
with the internal runner. This was due to a bug where the testrunner would incorrectly try
to install the exopackage files even when they were not available because the internal installation
had already taken place.

This diff adds logic to ensure we don't try to exo-install from the testrunner when running tests
internally.

Reviewed By: jtorkkola

fbshipit-source-id: 14c0cab1ac