[test fix] Fix issue where TravisCI sometimes does not recognize symlinks
This integration test works by checking what builds were invoked, and
their UUIDs. For some reason on TravisCI (only), one of the two last_*
links in buck-out/log are reporting 'true' for Files.isSymbolicLink,
and it would cause last_querycommand and the UUID based dir to show up
in the logs. Looking at the code in GlobalStateManager, this should
work, but unblock travisci for now, since this test isn't failing
anywhere else

Fixes #2229