Fix query --output-attributes to work with configurable non-concatenable attributes
Summary:
`JsonTypeConcatenatingCoercerFactory` eagerly failing for types that don't support concatenation. It should only fail when such are actually used in concatenation.

The fix is to introduce `SingleElementJsonTypeConcatenatingCoercer` that would return `null` in `concat` to indicate that it doesn't support concatenation. The logic that resolves select handles the cases when there is only one element, nothing else needs to be done here.

Reviewed By: jtorkkola

fbshipit-source-id: 620cb0481e