Rewrite HybridLocalStrategy
Summary:
There's currently a race here where we might not steal work that we
should. In general, it's hard to reason about the correctness of this
code.

This rewrites it in three major ways:

1. A single thread executor service is used for scheduling so we don't
need to think about scheduling races.
2. The JobLimiters are removed and replaced with explicit semaphores so that
the behavior is more obvious.
3. Tracking of the stage is removed so that we don't need to reason
about the correctness of that.

A test is added that just schedules a ton of no-op things with a remote
strategy that never completes and expects them to all get stolen. This
consistently saw failures in the previous code and ran 100000s of times
without failures in the new code (and locked up my Intellij after...).

Reviewed By: rajyengi

fbshipit-source-id: dd4b2afd49