Introduce BuildTargetPatternData and reimplement target pattern parsing
Summary:
BuildTargetPatternData will be used as a part of the key on early transformation stages of Graph Engine implementation of parsing computations. It is just a parsed representation of build target search pattern like "cell//some/path/to/...".

This is similar to BuildTargetPattern / BuildTargetPatternParser but not quite the same.

BuildTargetPattern is an interface having some polymorphic implementations which hide data. This is not appropriate for Graph Engine as it needs to know data and operate on data and not abstractions. So BuildTargetPatternData is pure data class free from abstractions or processing logic, as well as free of environment state, like absolute cell path, which current BuildTargetPattern implementations are unfortunate to have.

As for BuildTargetPatternParser, its implementation seems overcomplicated for simple things that it does, and may be overgeneralized too. I do not see how it can be reused to parse to BuildTargetPatternParserData without introducing more complexity to the code. I've come up with a separate more simple implementation in BuildTargetPatternDataParser with even better selection of unit tests.

BuildTargetPattern / BuildTargetPatternParser are currently used in different places and it is really hard to change them in place, so I am coming with this standalone implementation. Eventually BuildTargetPatternData / BuildTargetPatternDataParser will replace BuildTargetPattern / BuildTargetPatternParser (and satellites classes too).

Reviewed By: bobyangyf

fbshipit-source-id: fd97b89250