MungingDebugPathSanitizer is not needed anymore
Summary:
MungingDebugPathSanitizer was introduced some time ago because PrefixMapDebugPathSanitizer (using -fdebug-prefix-map) did not work for assembler with clang. The root cause was that clang did not pass this particular option to the assembler backend (gcc always did). This feature/bug/peculiarity was fixed in llvm (https://reviews.llvm.org/D48988) and finally the latest xcode 10.2 also got this update (ndk tools got it a while ago). So, there is no real need to keep this tricky and low-level MungingDebugPathSanitizer.

Also removing CxxPlatform.assemblerDebugPathSanitizer - since it is always the same as compilerDebugSanitizer now.

Reviewed By: cjhopman

fbshipit-source-id: 3225180898