Make implib from dll linking accessible output
Summary:
When creating a .dll for shipping, it is also required that we
ship the import lib for that dll, which is automatically created by
link.exe (and used by buck for linking dependents), but not exposed
as an output for genrules to copy. This diff makes it copyable
by a supported mechanism and not by hacks.

Reviewed By: ilya-klyuchnikov

fbshipit-source-id: f8c810906d