Fallback to num threads when local_jobs config is unset
Summary:
We're currently relying on a config value of `local_jobs=16` checked into the buckconfig in the repo.

Ideally we want a different number here for servers/laptops. Our original plan had been to use the experiments platform to roll different configs (ex: local_jobs=4 for Mac OS). However based on the order that the configs are read, this config would always get trampled by the default set in buckconfig in the repo. https://buckbuild.com/files-and-dirs/buckconfig.html#config-precedence

So allowing for local_jobs to be unset allows us to remove the repo level default, and tune the value with the experiments platform or using the num threads config we're using else where (this seems to be configured by thread_core_ratio which will scale appropriately based on the machine's cores).

Reviewed By: cjhopman

fbshipit-source-id: 991940a644