Use absolute paths for DERIVED_FILE_DIR (#2185)
Summary:
If a `.mlmodel` is added to an Xcode target where that module's `DERIVED_FILE_DIR` is set to a relative path, Xcode will crash.

This PR sets `DERIVED_FILE_DIR` to be an absolute path, which allows `.mlmodel`s to be included in generated Xcode projects without Xcode crashing.
Pull Request resolved: https://github.com/facebook/buck/pull/2185

Reviewed By: styurin

Pulled By: styurin

fbshipit-source-id: 9c7a8bfda0