Isolate testlib repository from the rest of Buck repository
Summary:
Test lib was in weird state, it was used as a cell in tests at the same time being a subgraph in Buck repository. This causes problems like inability to use local target names in `deps` (targets in `deps` must be present in both the test lib and Buck repository).

This change converts the folder to a proper cell.

Reviewed By: jtorkkola

fbshipit-source-id: 0a8cd7a0ea