Introduce MultitenantQueryEnvironment.
Summary:
This is a pretty big milestone as we can now do the equivalent of `buck query`
over a "graph" backed by `com.facebook.buck.multitenant.service.Index`!

This diff introduces `MultitenantQueryEnvironment`, which is an implementation
of `QueryEnvironment` backed by an `Index`, though most of the methods are
unimplemented. Its `getFunctions()` method returns only two functions:
`DepsFunction` and `DepsFunction.FirstOrderDepsFunction`.

Note that we only had to implement two methods to get `deps()` working
end-to-end: `getTargetEvaluator()` and `getFwdDeps()`.

The playbook going forward will be to fill out the other methods of
`MultitenantQueryEnvironment` so that we can add support for all of the
functions (returned by `getFunctions()` that `BuckQueryEnvironment` does.
Hopefully this will be fairly straightforward!

Reviewed By: sbalabanov

fbshipit-source-id: 2782e2b501