Move some classes into com.facebook.buck.core.model.targetgraph.raw
Summary:
We want the multitenant package to be able to use
`RawTargetNodeWithDeps` and `RawTargetNodeWithDepsPackage`
without having to pull in all of the transitive deps of
`com.facebook.buck.core.model.targetgraph`.

As we already moved `RawTargetNode` into its own package
(`com.facebook.buck.core.model.targetgraph.raw`) for this purpose,
this seems like the right move.

Reviewed By: bobyangyf

fbshipit-source-id: d1ece369f8