Make it possible to get a RawTargetNodeWithDeps from Index
Summary:
It is important to be able to retrieve a `RawTargetNodeWithDeps` from
the `Index` so the result can be printed with its attributes
(e.g., `buck query --output-attributes`).

This introduces `Index.getTargetNodes()` for fetching a number or target
nodes for a single commit. We also provide `Index.getTargetNode()` for
convenience, but the batch method should be preferred so the read lock
is only acquired once.

Note this should make a server fairly easy to implement:

```
/// It is up to the caller how to present the return value to the user: print as json, dot, etc.
/// Note that none of the values in the returned list should be null, so can throw for those?
fun doQuery(commit: Commit, query: String): List<RawTargetNodeWithDeps?> {
    /// Assume index is in scope and held in memory by the server:
    val env = MultitenantQueryEnvironment(index, commit)
    val unconfiguredTargets = env.evaluateQuery(query)
    return index.getTargetNodes(commit, unconfiguredTargets.toList())
}
```

Reviewed By: bobyangyf, sbalabanov

fbshipit-source-id: 341c5f3173