Move populateIndexFromStream from src/ to test/
Summary:
While here, I also changed the format of the JSON to be more similar to
what `buck query` spits out when `--output-attributes` is used. In
particular, it is now possible to specify the type of the build rule
via the `buck.type` property.

Because `AbstractRuleType` is package-private, I cannot seem to access
`AbstractRuleType.Kind` from Kotlin via `RuleType` as you can in Java.
To that end, I had to create a `RuleTypeFactory` in Java that is
explicitly for use with `IndexBuilder.kt`.

I also got rid of `//test/com/facebook/buck/multitenant/service:testutil`.
There was code specific to `MapDiffTest` that I inlined into that test
and everything else of value is now in `IndexBuilder`.

Reviewed By: sbalabanov

fbshipit-source-id: 8b40bdd36c